LPS-203604 Merge portal-impl LayoutModelListener logic into layout-impl LayoutModelListener
