+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/model/listener/LayoutModelListener.java
@@ -6,18 +6,25 @@
 package com.liferay.layout.internal.model.listener;
 
 import com.liferay.client.extension.service.ClientExtensionEntryRelLocalService;
+import com.liferay.exportimport.kernel.staging.LayoutStagingUtil;
 import com.liferay.friendly.url.service.FriendlyURLEntryLocalService;
 import com.liferay.layout.friendly.url.LayoutFriendlyURLEntryHelper;
 import com.liferay.layout.model.LayoutLocalization;
 import com.liferay.layout.service.LayoutClassedModelUsageLocalService;
 import com.liferay.layout.service.LayoutLocalizationLocalService;
 import com.liferay.portal.kernel.exception.ModelListenerException;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.language.Language;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.ModelListener;
+import com.liferay.portal.kernel.service.LayoutRevisionLocalServiceUtil;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.servlet.filters.cache.CacheUtil;
 
 import java.util.Locale;
 
@@ -30,6 +37,11 @@ import org.osgi.service.component.annotations.Reference;
  */
 @Component(service = ModelListener.class)
 public class LayoutModelListener extends BaseModelListener<Layout> {
+
+	@Override
+	public void onAfterCreate(Layout layout) {
+		clearCache(layout);
+	}
 
 	@Override
 	public void onAfterRemove(Layout layout) {
@@ -67,13 +79,59 @@ public class LayoutModelListener extends BaseModelListener<Layout> {
 					layoutLocalization);
 			}
 		}
+
+		clearCache(layout);
+	}
+
+	@Override
+	public void onAfterUpdate(Layout originalLayout, Layout layout) {
+		clearCache(layout);
 	}
 
 	@Override
 	public void onBeforeRemove(Layout layout) throws ModelListenerException {
 		_layoutClassedModelUsageLocalService.
 			deleteLayoutClassedModelUsagesByPlid(layout.getPlid());
+
+		try {
+			if ((layout == null) ||
+				!LayoutStagingUtil.isBranchingLayout(layout)) {
+
+				return;
+			}
+
+			LayoutRevisionLocalServiceUtil.deleteLayoutLayoutRevisions(
+				layout.getPlid());
+		}
+		catch (IllegalStateException illegalStateException) {
+			if (_log.isDebugEnabled()) {
+				_log.debug(illegalStateException);
+			}
+
+			// This is only needed because of LayoutPersistenceTest but should
+			// never happen in a deployed environment
+
+		}
+		catch (PortalException portalException) {
+			throw new ModelListenerException(portalException);
+		}
+		catch (SystemException systemException) {
+			throw new ModelListenerException(systemException);
+		}
 	}
+
+	protected void clearCache(Layout layout) {
+		if (layout == null) {
+			return;
+		}
+
+		if (!layout.isPrivateLayout()) {
+			CacheUtil.clearCache(layout.getCompanyId());
+		}
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		LayoutModelListener.class);
 
 	@Reference
 	private ClientExtensionEntryRelLocalService
